Fixed sign-versus-unsigned warning which afflicted the DOS port.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@1560 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csutil/vfs.cpp
@@ -323,7 +323,7 @@ DiskFile::DiskFile (int Mode, VfsNode *ParentNode, int RIndex,
     if (fseek (file, 0, SEEK_END))
       CheckError ();
     Size = ftell (file);
-    if (Size == -1)
+    if (Size == (size_t)-1)
     {
       Size = 0;
       CheckError ();
